2026_pj1_release/part2/unpytorched/requirements.txt
numpy
matplotlib
tqdm
# Optional: only required for --device gpu
cupy-cuda12x
